+++ src/TitleDetails/AddPieceModal/AddPieceModalContainer.test.js
@@ -3,10 +3,13 @@ import { render, fireEvent } from '@testing-library/react';
 import { IntlProvider } from 'react-intl';
 import { MemoryRouter } from 'react-router-dom';
 
-import '@folio/stripes-acq-components/test/jest/__mock__';
-
 import { PIECE_STATUS } from '@folio/stripes-acq-components';
 import AddPieceModalContainer from './AddPieceModalContainer';
+
+jest.mock('@folio/stripes-acq-components', () => ({
+  ...jest.requireActual('@folio/stripes-acq-components'),
+  FieldInventory: jest.fn(() => 'FieldInventory'),
+}));
 
 const renderAddPieceModalContainer = (
   close,
@@ -58,8 +61,6 @@ describe('AddPieceModalContainer', () => {
     expect(getByLabelText('ui-receiving.piece.caption')).toBeDefined();
     expect(getByText('ui-receiving.piece.format')).toBeDefined();
     expect(getByText('ui-receiving.piece.receiptDate')).toBeDefined();
-    expect(getByText('ui-receiving.piece.location')).toBeDefined();
-    expect(getByLabelText('ui-receiving.piece.location')).toBeDefined();
     expect(queryByText('ui-receiving.piece.actions.quickReceive')).toBeTruthy();
     fireEvent.input(getByLabelText('ui-receiving.piece.caption'));
   });
@@ -78,7 +79,7 @@ describe('AddPieceModalContainer', () => {
     const locations = [{ name: 'Location', id: '001' }];
     const locationIds = ['001'];
 
-    const { getByLabelText, queryByText, queryByLabelText, getByText } = renderAddPieceModalContainer(
+    const { getByLabelText, queryByText, getByText } = renderAddPieceModalContainer(
       close,
       onSubmit,
       piece,
@@ -93,6 +94,5 @@ describe('AddPieceModalContainer', () => {
     expect(getByText('stripes-acq-components.piece.pieceFormat.physical')).toBeDefined();
     expect(getByLabelText('ui-receiving.piece.receiptDate').disabled).toBeFalsy();
     expect(queryByText('ui-receiving.piece.actions.quickReceive')).toBeFalsy();
-    expect(queryByLabelText('ui-receiving.piece.location')).toBeFalsy();
   });
 });