+++ src/TitleReceive/TitleReceiveList.js
@@ -14,7 +14,7 @@ import {
   TextField,
 } from '@folio/stripes/components';
 import {
-  FieldLocationFinal,
+  FieldInventory,
   getItemStatusLabel,
   PIECE_FORMAT_LABELS,
 } from '@folio/stripes-acq-components';
@@ -100,12 +100,18 @@ export const TitleReceiveList = ({
           const locationIds = locationId ? [...new Set([...poLineLocationIds, locationId])] : poLineLocationIds;
 
           return (
-            <FieldLocationFinal
+            <FieldInventory
+              instanceId={instanceId}
+              locationIds={locationIds}
+              locations={locations}
+
+              labelless
               locationLookupLabel={<FormattedMessage id="ui-receiving.piece.locationLookup" />}
-              prepopulatedLocationsIds={locationIds}
-              locationsForDict={locations}
-              name={`${field}[${record.rowIndex}].locationId`}
-              onChange={({ id }) => selectLocation(id, `${field}[${record.rowIndex}].locationId`)}
+
+              holdingName={`${field}[${record.rowIndex}].holdingId`}
+              locationName={`${field}[${record.rowIndex}].locationId`}
+
+              onChange={selectLocation}
             />
           );
         },