+++ src/common/utils/index.js
@@ -1,4 +1,3 @@
-export * from './fetchLocations';
 export * from './getDehydratedPiece';
 export * from './getHydratedPieces';
 export * from './handleCommonErrors';